Support Null in JSON parameter
i had JSON string like this : {"1":"Man";"0":"Woman";"NULL":"Unset"}
my column was nullable Integer. but when data passed to it "NULL" is not integer occurred. 
null is true in JSON not "NULL" or "null" (-null- without double quotation is truly null) but we cant have key with null value. we create fake null with 'NULL' and then convert it to null in middleware.